refactor(core): rename ChatModelOpenAI to OpenAIChatModel
