+++ docs/apis/ai-agent-api.md
@@ -122,9 +122,9 @@ type AIAgentToolChoice = "auto" | "none" | "required" | "router" | Agent;
 ### Creating a Basic AIAgent
 
 ```typescript
-import { AIAgent, ChatModelOpenAI } from "@aigne/core-next";
+import { AIAgent, OpenAIChatModel } from "@aigne/core-next";
 
-const model = new ChatModelOpenAI({
+const model = new OpenAIChatModel({
   apiKey: process.env.OPENAI_API_KEY,
   model: "gpt-4"
 });
@@ -143,9 +143,9 @@ console.log(output.text); // Output of AI's response
 ### Using AIAgent with Tools
 
 ```typescript
-import { AIAgent, FunctionAgent, ChatModelOpenAI } from "@aigne/core-next";
+import { AIAgent, FunctionAgent, OpenAIChatModel } from "@aigne/core-next";
 
-const model = new ChatModelOpenAI({
+const model = new OpenAIChatModel({
   apiKey: process.env.OPENAI_API_KEY,
   model: "gpt-4"
 });
@@ -184,9 +184,9 @@ console.log(output.text); // "The weather in Beijing today is sunny with a tempe
 ### Using AIAgent in Router Mode
 
 ```typescript
-import { AIAgent, FunctionAgent, ChatModelOpenAI } from "@aigne/core-next";
+import { AIAgent, FunctionAgent, OpenAIChatModel } from "@aigne/core-next";
 
-const model = new ChatModelOpenAI({
+const model = new OpenAIChatModel({
   apiKey: process.env.OPENAI_API_KEY,
   model: "gpt-4"
 });
