+++ docs/apis/ai-agent-api.zh.md
@@ -122,9 +122,9 @@ type AIAgentToolChoice = "auto" | "none" | "required" | "router" | Agent;
 ### 创建基本的 AIAgent
 
 ```typescript
-import { AIAgent, ChatModelOpenAI } from "@aigne/core-next";
+import { AIAgent, OpenAIChatModel } from "@aigne/core-next";
 
-const model = new ChatModelOpenAI({
+const model = new OpenAIChatModel({
   apiKey: process.env.OPENAI_API_KEY,
   model: "gpt-4"
 });
@@ -143,9 +143,9 @@ console.log(output.text); // 输出 AI 的回复
 ### 使用工具的 AIAgent
 
 ```typescript
-import { AIAgent, FunctionAgent, ChatModelOpenAI } from "@aigne/core-next";
+import { AIAgent, FunctionAgent, OpenAIChatModel } from "@aigne/core-next";
 
-const model = new ChatModelOpenAI({
+const model = new OpenAIChatModel({
   apiKey: process.env.OPENAI_API_KEY,
   model: "gpt-4"
 });
@@ -184,9 +184,9 @@ console.log(output.text); // "北京今天天气晴朗，温度24°C，湿度60%
 ### 使用路由模式的 AIAgent
 
 ```typescript
-import { AIAgent, FunctionAgent, ChatModelOpenAI } from "@aigne/core-next";
+import { AIAgent, FunctionAgent, OpenAIChatModel } from "@aigne/core-next";
 
-const model = new ChatModelOpenAI({
+const model = new OpenAIChatModel({
   apiKey: process.env.OPENAI_API_KEY,
   model: "gpt-4"
 });
