+++ docs/apis/execution-engine-api.md
@@ -270,9 +270,9 @@ Predefined special topics:
 ### Basic Usage
 
 ```typescript
-import { ExecutionEngine, AIAgent, ChatModelOpenAI } from "@aigne/core-next";
+import { ExecutionEngine, AIAgent, OpenAIChatModel } from "@aigne/core-next";
 
-const model = new ChatModelOpenAI({
+const model = new OpenAIChatModel({
   apiKey: process.env.OPENAI_API_KEY,
   model: "gpt-4"
 });
@@ -308,9 +308,9 @@ await engine.shutdown();
 ### Sequential Execution of Multiple Agents
 
 ```typescript
-import { ExecutionEngine, AIAgent, FunctionAgent, sequential, ChatModelOpenAI } from "@aigne/core-next";
+import { ExecutionEngine, AIAgent, FunctionAgent, sequential, OpenAIChatModel } from "@aigne/core-next";
 
-const model = new ChatModelOpenAI({
+const model = new OpenAIChatModel({
   apiKey: process.env.OPENAI_API_KEY,
   model: "gpt-4"
 });
@@ -353,9 +353,9 @@ console.log(result);
 ### Parallel Execution of Multiple Agents
 
 ```typescript
-import { ExecutionEngine, AIAgent, parallel, ChatModelOpenAI } from "@aigne/core-next";
+import { ExecutionEngine, AIAgent, parallel, OpenAIChatModel } from "@aigne/core-next";
 
-const model = new ChatModelOpenAI({
+const model = new OpenAIChatModel({
   apiKey: process.env.OPENAI_API_KEY,
   model: "gpt-4"
 });
@@ -392,9 +392,9 @@ console.log("Story:", result.story);
 ### Using Publish-Subscribe Pattern
 
 ```typescript
-import { ExecutionEngine, AIAgent, FunctionAgent, ChatModelOpenAI } from "@aigne/core-next";
+import { ExecutionEngine, AIAgent, FunctionAgent, OpenAIChatModel } from "@aigne/core-next";
 
-const model = new ChatModelOpenAI({
+const model = new OpenAIChatModel({
   apiKey: process.env.OPENAI_API_KEY,
   model: "gpt-4"
 });
