+++ docs/apis/execution-engine-api.zh.md
@@ -270,9 +270,9 @@ interface UserAgentOptions<I extends Message = Message, O extends Message = Mess
 ### 基本用法
 
 ```typescript
-import { ExecutionEngine, AIAgent, ChatModelOpenAI } from "@aigne/core-next";
+import { ExecutionEngine, AIAgent, OpenAIChatModel } from "@aigne/core-next";
 
-const model = new ChatModelOpenAI({
+const model = new OpenAIChatModel({
   apiKey: process.env.OPENAI_API_KEY,
   model: "gpt-4"
 });
@@ -308,9 +308,9 @@ await engine.shutdown();
 ### 顺序执行多个 Agent
 
 ```typescript
-import { ExecutionEngine, AIAgent, FunctionAgent, sequential, ChatModelOpenAI } from "@aigne/core-next";
+import { ExecutionEngine, AIAgent, FunctionAgent, sequential, OpenAIChatModel } from "@aigne/core-next";
 
-const model = new ChatModelOpenAI({
+const model = new OpenAIChatModel({
   apiKey: process.env.OPENAI_API_KEY,
   model: "gpt-4"
 });
@@ -353,9 +353,9 @@ console.log(result);
 ### 并行执行多个 Agent
 
 ```typescript
-import { ExecutionEngine, AIAgent, parallel, ChatModelOpenAI } from "@aigne/core-next";
+import { ExecutionEngine, AIAgent, parallel, OpenAIChatModel } from "@aigne/core-next";
 
-const model = new ChatModelOpenAI({
+const model = new OpenAIChatModel({
   apiKey: process.env.OPENAI_API_KEY,
   model: "gpt-4"
 });
@@ -392,9 +392,9 @@ console.log("故事:", result.story);
 ### 使用发布-订阅模式
 
 ```typescript
-import { ExecutionEngine, AIAgent, FunctionAgent, ChatModelOpenAI } from "@aigne/core-next";
+import { ExecutionEngine, AIAgent, FunctionAgent, OpenAIChatModel } from "@aigne/core-next";
 
-const model = new ChatModelOpenAI({
+const model = new OpenAIChatModel({
   apiKey: process.env.OPENAI_API_KEY,
   model: "gpt-4"
 });
