+++ docs/cookbook.zh.md
@@ -77,7 +77,7 @@ const engine = new ExecutionEngine({ model });
 **示例**:
 
 ```typescript
-import { AIAgent, ChatModelOpenAI, ExecutionEngine, FunctionAgent } from "@aigne/core-next";
+import { AIAgent, OpenAIChatModel, ExecutionEngine, FunctionAgent } from "@aigne/core-next";
 import { z } from "zod";
 
 // 创建JavaScript沙箱
@@ -123,7 +123,7 @@ console.log(result);
 **示例**:
 
 ```typescript
-import { AIAgent, ChatModelOpenAI, ExecutionEngine } from "@aigne/core-next";
+import { AIAgent, OpenAIChatModel, ExecutionEngine } from "@aigne/core-next";
 
 // 概念提取Agent
 const conceptExtractor = AIAgent.from({
@@ -192,7 +192,7 @@ console.log(result);
 **示例**:
 
 ```typescript
-import { AIAgent, ChatModelOpenAI, ExecutionEngine, parallel } from "@aigne/core-next";
+import { AIAgent, OpenAIChatModel, ExecutionEngine, parallel } from "@aigne/core-next";
 
 // 功能提取Agent
 const featureExtractor = AIAgent.from({
@@ -240,7 +240,7 @@ console.log(result);
 ```typescript
 import {
   AIAgent,
-  ChatModelOpenAI,
+  OpenAIChatModel,
   ExecutionEngine,
   UserInputTopic,
   UserOutputTopic,
@@ -321,7 +321,7 @@ console.log(result);
 **示例**:
 
 ```typescript
-import { AIAgent, ChatModelOpenAI, ExecutionEngine } from "@aigne/core-next";
+import { AIAgent, OpenAIChatModel, ExecutionEngine } from "@aigne/core-next";
 
 // 转交给Agent B的函数
 function transfer_to_b() {
@@ -370,7 +370,7 @@ console.log(result2);
 **示例**:
 
 ```typescript
-import { AIAgent, ChatModelOpenAI, ExecutionEngine } from "@aigne/core-next";
+import { AIAgent, OpenAIChatModel, ExecutionEngine } from "@aigne/core-next";
 
 // 产品支持Agent
 const productSupport = AIAgent.from({
@@ -447,7 +447,7 @@ console.log(result3);
 
 ```typescript
 import { OrchestratorAgent } from "@aigne/agent-library";
-import { AIAgent, ChatModelOpenAI, ExecutionEngine, MCPAgent } from "@aigne/core-next";
+import { AIAgent, OpenAIChatModel, ExecutionEngine, MCPAgent } from "@aigne/core-next";
 
 // 创建各专业Agent
 const puppeteer = await MCPAgent.from({
@@ -554,7 +554,7 @@ Puppeteer MCP服务器允许AIGNE Framework访问和操作网页内容。
 ```typescript
 import {
   AIAgent,
-  ChatModelOpenAI,
+  OpenAIChatModel,
   ExecutionEngine,
   MCPAgent,
 } from "@aigne/core-next";
@@ -609,7 +609,7 @@ SQLite MCP服务器允许AIGNE Framework与SQLite数据库交互。
 import { join } from "node:path";
 import {
   AIAgent,
-  ChatModelOpenAI,
+  OpenAIChatModel,
   ExecutionEngine,
   MCPAgent,
 } from "@aigne/core-next";
