+++ examples/mcp-server-sqlite/index.ts
@@ -4,9 +4,9 @@ import assert from "node:assert";
 import { join } from "node:path";
 import {
   AIAgent,
-  ChatModelOpenAI,
   ExecutionEngine,
   MCPAgent,
+  OpenAIChatModel,
   PromptBuilder,
   getMessage,
   logger,
@@ -18,7 +18,7 @@ assert(OPENAI_API_KEY, "Please set the OPENAI_API_KEY environment variable");
 
 logger.enable(`aigne:mcp,${process.env.DEBUG}`);
 
-const model = new ChatModelOpenAI({
+const model = new OpenAIChatModel({
   apiKey: OPENAI_API_KEY,
 });
