+++ examples/workflow-code-execution/index.ts
@@ -3,9 +3,9 @@
 import assert from "node:assert";
 import {
   AIAgent,
-  ChatModelOpenAI,
   ExecutionEngine,
   FunctionAgent,
+  OpenAIChatModel,
   runChatLoopInTerminal,
 } from "@aigne/core-next";
 import { z } from "zod";
@@ -13,7 +13,7 @@ import { z } from "zod";
 const { OPENAI_API_KEY } = process.env;
 assert(OPENAI_API_KEY, "Please set the OPENAI_API_KEY environment variable");
 
-const model = new ChatModelOpenAI({
+const model = new OpenAIChatModel({
   apiKey: OPENAI_API_KEY,
 });
