+++ examples/workflow-code-execution/usage.ts
@@ -1,11 +1,11 @@
 import assert from "node:assert";
-import { AIAgent, ChatModelOpenAI, ExecutionEngine, FunctionAgent } from "@aigne/core-next";
+import { AIAgent, ExecutionEngine, FunctionAgent, OpenAIChatModel } from "@aigne/core-next";
 import { z } from "zod";
 
 const { OPENAI_API_KEY } = process.env;
 assert(OPENAI_API_KEY, "Please set the OPENAI_API_KEY environment variable");
 
-const model = new ChatModelOpenAI({
+const model = new OpenAIChatModel({
   apiKey: OPENAI_API_KEY,
 });
