+++ examples/workflow-concurrency/index.ts
@@ -3,8 +3,8 @@
 import assert from "node:assert";
 import {
   AIAgent,
-  ChatModelOpenAI,
   ExecutionEngine,
+  OpenAIChatModel,
   parallel,
   runChatLoopInTerminal,
 } from "@aigne/core-next";
@@ -12,7 +12,7 @@ import {
 const { OPENAI_API_KEY } = process.env;
 assert(OPENAI_API_KEY, "Please set the OPENAI_API_KEY environment variable");
 
-const model = new ChatModelOpenAI({
+const model = new OpenAIChatModel({
   apiKey: OPENAI_API_KEY,
 });
