+++ examples/workflow-sequential/README.md
@@ -72,12 +72,12 @@ The following example demonstrates how to build a sequential workflow:
 
 ```typescript
 import assert from "node:assert";
-import { AIAgent, ChatModelOpenAI, ExecutionEngine, sequential } from "@aigne/core-next";
+import { AIAgent, OpenAIChatModel, ExecutionEngine, sequential } from "@aigne/core-next";
 
 const { OPENAI_API_KEY } = process.env;
 assert(OPENAI_API_KEY, "Please set the OPENAI_API_KEY environment variable");
 
-const model = new ChatModelOpenAI({
+const model = new OpenAIChatModel({
   apiKey: OPENAI_API_KEY,
 });
