+++ packages/core/src/index.ts
@@ -5,9 +5,9 @@ export * from "./agents/memory.js";
 export * from "./agents/types.js";
 export * from "./agents/user-agent.js";
 export * from "./execution-engine/index.js";
-export * from "./models/chat-claude.js";
-export * from "./models/chat-openai.js";
-export * from "./models/chat.js";
+export * from "./models/chat-model.js";
+export * from "./models/claude-chat-model.js";
+export * from "./models/openai-chat-model.js";
 export * from "./prompt/prompt-builder.js";
 export * from "./prompt/template.js";
 export * from "./utils/logger.js";
