+++ packages/core/src/models/chat-model.ts
