+++ packages/core/test/models/chat-openai.test.ts
@@ -4,7 +4,7 @@ import { join } from "node:path";
 import {
   AgentMessageTemplate,
   ChatMessagesTemplate,
-  ChatModelOpenAI,
+  OpenAIChatModel,
   SystemMessageTemplate,
   ToolMessageTemplate,
   UserMessageTemplate,
@@ -14,8 +14,8 @@ import type { APIPromise } from "openai/core";
 import type { ChatCompletion, ChatCompletionChunk } from "openai/resources";
 import type { Stream } from "openai/streaming";
 
-test("ChatModelOpenAI.call", async () => {
-  const model = new ChatModelOpenAI({
+test("OpenAIChatModel.call", async () => {
+  const model = new OpenAIChatModel({
     apiKey: "YOUR_API_KEY",
     model: "gpt-4o-mini",
   });
